MVP for Nvidia and AMD Gpu telemetry complete. Improve with better granularity in the future. Still needs to output to JSON for future DTO retrieval and creation by the QT frontend when we make it
