+++ System_Info/hardware/GPU/NVIDIA/NvidiaPool.h
@@ -1,0 +1,20 @@
+#pragma once
+
+#include <windows.h>
+#include "..\..\..\..\third_party\nvapi\nvapi.h"
+#include "NvidiaLiveGPUMetrics.h"
+#include "..\..\IHardwareDevicePool.h"
+#include <vector>
+
+class NvidiaPool : public IHardwareDevicePool
+{
+public:
+	NvidiaPool();
+	~NvidiaPool() override;
+
+	void enumerateDevices() override;
+
+private:
+	std::vector<std::unique_ptr<NvidiaLiveGPUMetrics>> nvidia_adapters;
+};
+
