--- System_Info/hardware/GPU/NvidiaPool.cpp
@@ -1,1 +1,0 @@
-#include "NvidiaPool.h"
