--- System_Info/hardware/GPU/NvidiaPool.h
@@ -1,6 +1,0 @@
-#pragma once
-class NvidiaPool
-{
-
-};
-
