+++ System_Info/hardware/HardwareDevice.cpp
@@ -3,6 +3,8 @@
 #include <iostream>
 
 void A_HardwareDevice::outputMetrics() const {
+    OUTPUT_HEADER("Live GPU Metrics")
+
     std::cout << "-- " << name << " --\n";
     for (const auto& sensor : dev_sensors) {
         std::cout << std::left << std::setw(32) << sensor->getName()
